refact: consolidate numeric input handling and adjust price utilities into shared utils
